Create static screens views
