--- js/screen-intro.js
@@ -1,23 +1,0 @@
-import createDomElementFromStringTemplate from './create-dom-element';
-import renderScreen from "./render-screen";
-import greetingScreenElement from "./screen-greeting";
-
-const introScreenTemplate = `
-<section class="intro">
-  <button class="intro__asterisk asterisk" type="button"><span class="visually-hidden">Продолжить</span>*</button>
-  <p class="intro__motto"><sup>*</sup> Это не фото. Это рисунок маслом нидерландского художника-фотореалиста Tjalf Sparnaay.</p>
-</section>
-`;
-
-const introScreenElement = createDomElementFromStringTemplate(introScreenTemplate);
-const introScreenNext = introScreenElement.querySelector(`.asterisk`);
-
-const onIntroScreenNextClick = (evt) => {
-  evt.preventDefault();
-
-  renderScreen(greetingScreenElement);
-};
-
-introScreenNext.addEventListener(`click`, onIntroScreenNextClick);
-
-export default introScreenElement;
